Rework `Resolver` to follow middleware pattern for name lookup.
