+++ guides/getting-started/readme.md
@@ -1,0 +1,16 @@
+# Getting Started
+
+This guide explains how to get started with the `async-dns` gem.
+
+## Installation
+
+Add the gem to your project:
+
+~~~ bash
+$ bundle add async-dns
+~~~
+
+## Usage
+
+### Client
+
