+++ lib/async/dns.rb
@@ -8,7 +8,7 @@ require 'async'
 require_relative 'dns/version'
 
 require_relative 'dns/server'
-require_relative 'dns/resolver'
+require_relative 'dns/client'
 require_relative 'dns/handler'
 
 # @namespace
